Looking through the old University code
Decided to go ahead and see what the code I wrote
for an intro to OOP class looked like from University.
Suprised that it compiles... and is basically readable.
Must not have been too bad a class. Though I have a lot
of use of shared_ptr there... probably not great.

Still... happy enough just to take a look back.